Calculator.py
changed the type of the input variable so that we can use decimal numbers for the calculation

+++ Calculator.py
@@ -55,8 +55,8 @@ def division(x, y):
 continue_calculation = "y"
 
 while continue_calculation == "y":
-    num1 = int(input("Enter the first number: "))
-    num2 = int(input("Enter the second number: "))
+    num1 = float(input("Enter the first number: "))
+    num2 = float(input("Enter the second number: "))
     operation = input("Enter the operation you want to perform (+, -, *, /): ")
 
     if operation == '+':
